backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
supabase<2.5.0
cloudinary==1.41.0
groq
Pillow==10.4.0
email-validator==2.3.0
